requirements.txt
absl-py==1.3.0
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
isort==5.10.1
kiwisolver==1.4.4
matplotlib==3.6.2
networkx==2.8.8
numpy==1.23.4
ortools==9.4.1874
packaging==21.3
Pillow==9.3.0
protobuf==4.21.9
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
